# internal/presentation/rest/handlers/dto.go
package handlers

// --- DTOs compartidos entre todos los handlers ---

// ErrorResponse es la respuesta estándar para errores controlados.
type ErrorResponse struct {
	Error string `json:"error" example:"invalid credentials"`
}

// MessageResponse es una respuesta genérica con mensaje informativo.
type MessageResponse struct {
	Message string `json:"message" example:"user registered successfully"`
}